% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_add_px.R
\name{add_px}
\alias{add_px}
\title{Add px}
\usage{
add_px(val)
}
\arguments{
\item{val}{a value of type character or which can be coerced to type character}
}
\value{
the value with " px" appended
}
\description{
Just adds " px" to the end of the input value
}
\examples{
add_px(10)
}
