% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_fonts.R
\name{get_fonts}
\alias{get_fonts}
\title{Get Fonts}
\usage{
get_fonts(css_list)
}
\arguments{
\item{css_list}{the list of css elements, where each element is a list
containing a vector of attributes and a vector of values}
}
\value{
the unique font families used
}
\description{
Extracts fonts from the css list
}
