% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_font_import.R
\name{make_font_import}
\alias{make_font_import}
\title{Make Font Import}
\usage{
make_font_import(fonts)
}
\arguments{
\item{fonts}{a vector of font names}
}
\value{
a string that can be used in a css file to import fonts
}
\description{
Turns a vector of fonts into the font import line.
}
