% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_make_mod_from_elem.R
\name{make_mod_ui_from_att}
\alias{make_mod_ui_from_att}
\title{Make mod UI from attent}
\usage{
make_mod_ui_from_att(att)
}
\arguments{
\item{att}{a character containing an HTML attent name}
}
\value{

}
\description{
Takes an attent name and returns the UI module function for that attent.
}
