% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_unreactify.R
\name{nested_unreactify}
\alias{nested_unreactify}
\title{Nested Unreactify}
\usage{
nested_unreactify(x)
}
\arguments{
\item{x}{a list of lists, where the lower-level list contains some reactive
elements.}
}
\value{
a list of lists where the lower-level list of reactives has been
evaluated
}
\description{
This function is useful for when you have a list of lists, and you want
to unreactify the elements of the second-level lists.
}
