% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_unreactify.R
\name{unreactify}
\alias{unreactify}
\title{Unreactify}
\usage{
unreactify(elem)
}
\arguments{
\item{elem}{}
}
\value{
returned value of eleme if reactive, else elem
}
\description{
Returns the returned value of a reactive if the input arg is a reactive,
or simply returns its current value if not. This is useful for handling
returned lists of values from modules where some elements are reactive
but others are not. It may also be useful in cases where you want a module
to have the flexibility to use a reactive in some instances but a static 
value in others.
}
\examples{
example_1 <- reactive({ return("tabby cats") })
example_2 <- "tuxedo cats"
unreactify(example_1)
unreactify(example_2)
}
